refactor(approval-gate): extract wire.rs from lib.rs
First step of breaking lib.rs (~4300 lines) into focused modules.
This commit lifts the wire-format types and small wire-shape helpers
into their own module — pure data shapes with no I/O, no iii-sdk
dependency, no async — so downstream workers that only need to
understand the approval-gate protocol can depend on a small surface.

Moved (lib.rs -> wire.rs):
- Denial (structured deny payload)
- IncomingCall + requires_approval()
- Decision (Allow | Deny(Denial))
- WireDecision (allow|deny wire enum)
- pending_key()
- extract_call()
- block_reply_for()

All public items stay re-exported from the crate root, so every
existing call site (handlers, tests, integration tests, downstream
crates) continues to work without import changes.

Pure move, no behavior change. 141 approval-gate tests pass;
turn-orchestrator + harness + session + providers + harness-tui +
shell all pass.
